Apply suggested changes
